+++ Assets/Scripts/SquareData.cs
@@ -6,12 +6,12 @@ public class SquareData
 {
     public Utils.Cell cell;
     public int id;
-    public long value;
+    public float value;
     
 
     public Vector2 Position => Utils.GridToPos(cell.Row, cell.Column);
 
-    public SquareData(Utils.Cell cell, int id, long value)
+    public SquareData(Utils.Cell cell, int id, float value)
     {
         this.cell = cell;
         this.id = id;
